[TEST] Consolidate Gluon tcgen05 coverage
