--- tests/end_to_end/test_gluon_async_copy_ops.py
@@ -1,107 +1,0 @@
-import torch
-from triton.experimental import gluon
-from triton.experimental.gluon import language as gl
-from triton.experimental.gluon.language.nvidia.ampere import async_copy as cp
-
-import triton_viz
-
-
-@gluon.jit
-def _async_copy_1d_kernel(in_ptr, out_ptr, xnumel, BLOCK: gl.constexpr):
-    pid = gl.program_id(0)
-    layout: gl.constexpr = gl.BlockedLayout([1], [32], [4], [0])
-    offsets = pid * BLOCK + gl.arange(0, BLOCK, layout=layout)
-    mask = offsets < xnumel
-    smem_layout: gl.constexpr = gl.SwizzledSharedLayout(
-        vec=1,
-        per_phase=1,
-        max_phase=1,
-        order=[0],
-    )
-    smem = gl.allocate_shared_memory(gl.float32, [BLOCK], layout=smem_layout)
-
-    cp.async_load(smem, in_ptr + offsets, mask=mask)
-    cp.commit_group()
-    cp.wait_group(0)
-    gl.store(out_ptr + offsets, smem.load(layout), mask=mask)
-
-
-@gluon.jit
-def _async_copy_elementwise_add_kernel(
-    a_ptr,
-    b_ptr,
-    out_ptr,
-    xnumel,
-    ynumel,
-    xstride_a,
-    ystride_a,
-    xstride_b,
-    ystride_b,
-    xstride_out,
-    ystride_out,
-    XBLOCK: gl.constexpr,
-    YBLOCK: gl.constexpr,
-    smem_layout: gl.constexpr,
-):
-    pid = gl.program_id(0)
-    layout: gl.constexpr = gl.BlockedLayout([1, 1], [1, 32], [1, 4], [1, 0])
-    xoffs = pid * XBLOCK + gl.arange(0, XBLOCK, gl.SliceLayout(1, layout))
-    yoffs = gl.arange(0, YBLOCK, gl.SliceLayout(0, layout))
-    mask = (xoffs < xnumel)[:, None] & (yoffs < ynumel)[None, :]
-    a_smem = gl.allocate_shared_memory(gl.float32, [XBLOCK, YBLOCK], smem_layout)
-    b_smem = gl.allocate_shared_memory(gl.float32, [XBLOCK, YBLOCK], smem_layout)
-
-    cp.async_load(
-        a_smem,
-        a_ptr + xstride_a * xoffs[:, None] + ystride_a * yoffs[None, :],
-        mask=mask,
-    )
-    cp.async_load(
-        b_smem,
-        b_ptr + xstride_b * xoffs[:, None] + ystride_b * yoffs[None, :],
-        mask=mask,
-    )
-    cp.commit_group()
-    cp.wait_group(0)
-    values = a_smem.load(layout) + b_smem.load(layout)
-    gl.store(
-        out_ptr + xstride_out * xoffs[:, None] + ystride_out * yoffs[None, :],
-        values,
-        mask=mask,
-    )
-
-
-def test_gluon_async_copy_runs_masked_1d_copy_on_cpu():
-    inp = torch.arange(40, dtype=torch.float32)
-    out = torch.full_like(inp, -1)
-    kernel = triton_viz.trace("tracer", frontend="gluon")(_async_copy_1d_kernel)
-
-    kernel[(1,)](inp, out, inp.numel(), 64, num_warps=4)
-
-    torch.testing.assert_close(out, inp, atol=0, rtol=0)
-
-
-def test_gluon_async_copy_runs_staged_elementwise_add_on_cpu():
-    a = torch.arange(24, dtype=torch.float32).reshape(4, 6)
-    b = 10 + a
-    out = torch.full_like(a, -1)
-    smem_layout = gl.SwizzledSharedLayout(vec=1, per_phase=1, max_phase=1, order=[1, 0])
-    kernel = triton_viz.trace("tracer", frontend="gluon")(
-        _async_copy_elementwise_add_kernel
-    )
-
-    kernel[(1,)](
-        a,
-        b,
-        out,
-        *a.shape,
-        *a.stride(),
-        *b.stride(),
-        *out.stride(),
-        8,
-        8,
-        smem_layout,
-        num_warps=4,
-    )
-
-    torch.testing.assert_close(out, a + b, atol=0, rtol=0)
